Made a few updates in reference to #6.
The config.ini files now include a STEAMDIRS variable and its values are stored in constants.STEAM_DIRS.

A new games.ini file has been added to store known Steam games that Source.Python can be linked to.

The games.ini values are then checked on the system for existence and if found are included in constants.available_games.

config.ini's RELEASEDIR and PYTHONEXE paths are now stored in constants with the rest of the paths.